+++ rog-types/src/error.rs
@@ -28,6 +28,7 @@ impl Error for AuraError {}
 #[derive(Debug)]
 pub enum GraphicsError {
     ParseVendor,
+    ParsePower,
 }
 
 impl fmt::Display for GraphicsError {
@@ -35,8 +36,33 @@ impl fmt::Display for GraphicsError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
             GraphicsError::ParseVendor => write!(f, "Could not parse vendor name"),
+            GraphicsError::ParsePower => write!(f, "Could not parse dGPU power status"),
         }
     }
 }
 
 impl Error for GraphicsError {}
+
+#[derive(Debug)]
+pub enum AnimeError {
+    InvalidBitmap,
+    Io(std::io::Error),
+}
+
+impl fmt::Display for AnimeError {
+    // This trait requires `fmt` with this exact signature.
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            AnimeError::InvalidBitmap => write!(f, "Bitmap is invalid"),
+            AnimeError::Io(e) => write!(f, "Could not open: {}", e),
+        }
+    }
+}
+
+impl Error for AnimeError {}
+
+impl From<std::io::Error> for AnimeError {
+    fn from(err: std::io::Error) -> Self {
+        AnimeError::Io(err)
+    }
+}