+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Display PNG images on AniMe with scale, position, angle, and brightness
+- AniMe display parts split out to individual crate in preparation for publishing
+  on crates.io
+
 # [3.3.0] - 2021-04-3
 ### Changed
 - Add ledmodes for G733QS